feat: read api key from stdin
If the api key is not set read via environment read it from stdin
without echoing to stdout.
